Update additional information
